Fix display of tool on multiframe image
Need to include the frame number in the imageId for the tool
to recognize when it is meant to display.

Also need to pass in the displaySets into the SR loader
since they haven't been set yet in the OHIF data structures.

This commit also cleans up some debug messages.